Added core project
The idea is to have core be in .net core in the future. But since I
have had bad experiences with mixing .net framework with .net core I
wont do so just yet. Will wait for the tech to mature.

The new structure encourages sepeartion between Models (in core) and
Views and ViewModels in the CustomPresenter project.