MainWindowViewModel unit tests added
Does not look pretty. Would be nice to not have to create files. But
at the same time modifying the logic around that so that we can write
tests that does not have to create files does not sound worth it.

Skipped open file dialog tests since they open a dialog and require
user input.